Implement Day09 Part2

+++ src/main/kotlin/aoc/day09/Player.kt
@@ -1,3 +1,3 @@
 package aoc.day09
 
-data class Player(val id: Int, var points: Int = 0)
+data class Player(val id: Int, var points: Long = 0)
